+++ km3pipe/utils/runtable.py
@@ -3,12 +3,13 @@
 Prints the run table for a given detector ID.
 
 Usage:
-    runtable [-n RUNS] [-s REGEX] DET_ID
+    runtable [options] DET_ID
     runtable (-h | --help)
     runtable --version
 
 Options:
     -h --help           Show this screen.
+    -c                  Compact view.
     -n RUNS             Number of runs.
     -s REGEX            Regular expression to filter the runsetup name/id.
     DET_ID              Detector ID (eg. D_ARCA001).
@@ -30,7 +31,7 @@ __status__ = "Development"
 log = kp.logger.get_logger(__name__)
 
 
-def runtable(det_id, n=5, sep='\t', regex=None):
+def runtable(det_id, n=5, compact=False, sep='\t', regex=None):
     """Print the run table of the last `n` runs for given detector"""
     db = kp.db.DBManager()
     df = db.run_table(det_id)
@@ -48,6 +49,9 @@ def runtable(det_id, n=5, sep='\t', regex=None):
     if n is not None:
         df = df.tail(n)
 
+    if compact:
+        df = df[['RUN', 'DATETIME', 'RUNSETUPNAME']]
+
     df.to_csv(sys.stdout, sep=sep)
 
 
@@ -60,4 +64,4 @@ def main():
     except TypeError:
         n = None
 
-    runtable(args['DET_ID'], n, regex=args['-s'])
+    runtable(args['DET_ID'], n, regex=args['-s'], compact=args['-c'])